Add -r, --show-routes option to show VIFs and routing table of running pimd.
This changeset improves usability a bit by adding the -r, --show-routes,
option to pimd.  Calling pimd with this option is the same thing as when
you manually send SIGUSR1 to a running daemon and then displays the dump
file containing active VIFs and multicast routes.  Quite handy.

Support for SIGUSR2 and SIGQUIT does not seem to be implemented, but I
have left some skeleton code commented out for the day that is fixed.