Fix issue #3 "pimd: FTBFS, compiler warning + -Werror"
This commit should fix the FTBFS issue reported in Debian bug 595584.  It
seems GCC on, e.g., PowerPC and S/390 is slightly more cranky than usual.
Complaining about "may be used uninitialized in this function" when there
really shouldn't be anything to be concerned with.  However, this commit
makes sure to initialize all stack variables properly before leaving off
to the rest of the code.

Also, a lot of cleanup and ansification.