Implement RFC 3021 support
RFC 3021 mandates that /31 IPv4 netmasks should be treated as valid and that a
subnet without broadcast and subnet addresses is specified. All broadcast
communication must use the limited broadcast address (255.255.255.255).

/31 subnets are commonly used as router interconnection links, so we treat them as
point-to-point interfaces.

Because these interfaces do not carry the IFF_POINTOPOINT flag, subsequent
checks are changed to look for the VIFF_POINT_TO_POINT vif flag.

Signed-off-by: Apollon Oikonomopoulos <apoikos@gmail.com>