Do not skip IGMPv3 group joins with a given source address
In IGMPv3 membership reports an "optional" source field exists, pimd
does not yet support PIM-SSM, so we simply ignore the source-specific
request and treat it like a regular IGMPv2 join.  This is not optimal,
but is more compatible than skipping the group record altogether.

Also, cleanup and adapt the new IGMPv3 code to the original coding
style.  Changing to *BSD style headers, like rest of pimd.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>