PMTU draft support for register message fragmentation on EMSGSIZE
When the DR tunnels all multicast (of a given group, or groups) in PIM
register messages, it must take the PMTU into account.  Forwarding a
full-size multicast frame will likely not work in any standard Ethernet
setup.

In pim_send_unicast(), used for PIM_REGISTER messages, the sendto()
syscall will receive EMSGSIZE when the fram is too big.  It is up to the
application to fragment, and encapsulate each fragment, when forwarding
it to the RP.  When the RP receives the fragments the kernel will make
sure to reassemble the fragments before sending the fram to pimd again.

What remains to be done here is first of all to add support for the
Linux specific IP_RECVERR sockopt and MSG_ERRQUEUE recvfrom() flag to
read out the ee_info field of the sock_extended_err struct, which is the
PMTU on EMSGSIZE.  Then use that value as the size of each fragment
sent.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>