Add fragmentation support for unicast message that are larger than the MTU of the output interface.
Assuming all ethernet interfaces, the encap added to register messages
will cause the size of the packet to exceed the MTU of the outgoing
interface if the original packet is max size.  Since the socket has
option IP_HDRINCL set, the kernel will not fragment the IP packet.  So
pimd must do it.