Add fragmentation support for PIM register messages
The encapsulation added by register messages to multicast frames will
sometimes cause the size of the resulting frame to exceed the MTU of the
outbound interface.  Since the socket has option IP_HDRINCL set, the
kernel will not fragment the IP packet, so it is up to pimd to do it.

When receiving a fragmented PIM register message the UNIX kernel will
reassemble the whole frame before sending it to userspace, so there is
no need for this functionality in pimd.

Signed-off-by: Michael Fine <mfine@cumulusnetworks.com>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>