Rename new IGMP settings: default_igmp_ to igmp_
This patch renames the new settings `default_igmp_query_interval` and
`default_igmp_querier_timeout` to the easier `igmp_query_interval` and
`igmp_querier_timeout`.  This is both easier to read and less confusing
to the user.  Default otherwise implicates that the values change at
runtime, after some initial peering or whatnot, this is true for some of
the PIM related setting but not the case for these IGMP settings.

Backwards compatible support in the `.conf` file parser has been kept
for the sake of those who are already using them.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>