Let user know of any possible firewall problems.
If PIM or IGMP subsystems fail to send a frame it could be due to the
firewall.  This patch adds checks for two common `errno` error codes
returned by the Linux and OpenBSD kernels which pimd will log with an
appropriate warning.

On Linux systems you get `EPERM`: “Operation not permitted”, and on
OpenBSD you likely get `EHOSTUNREACH`: “No route to host” all over the
logs, usually for the IGMP traffic that pimd generates.  On OpenBSD this
is due to `pf` by default filtering all frames with IP options set.

Support for the “Router Alert” IP option to IGMP frames was added in
760dd9d, for the v2.2.0 release, as is recommended in [RFC2113][1].

Add this to your OpenBSD `/etc/pf.conf` to allow IGMP frames to be sent
by pimd:

    # Enable IGMP with Router Alert IP option
    pass proto igmp allow-opts

[1]: https://tools.ietf.org/html/rfc2113

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>