igmp.c: Attempt to clean up #ifdef hell in accept_igmp()
This patch attempts to clean up the #ifdef hell that's been building up
in igmp.c for a while now.  Let's test this fix out on Linux, OpenBSD,
FreeBSD and NetBSD. Fingers crossed!

The current mess is due to the various UNIX flavors handling SOCK_RAW
and IP_HDRINCL a bit differently.  It seems though that NetBSD is the
only one (of the major players) that still is a bit different for the
others.  NetBSD adjusts ip_len with the IP header size before it
forwards the frame to the userspace RAW socket.  Not so raw after all.

Since we cannot rely, in any sane way, on ip_len being correct, this
patch replaces the IGMP packet length sanity check that until now has
used `ip->ip_len` in igmp.c, with `recvlen` that pim.c already use in
`accept_pim()`.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>