Clean up conditional build of netlink (Linux) and routingsock (BSD)
This patch attempts to clean up more legacy #ifdef hell which we no
longer need thanks to the configure script.

On Linux we continue to use the Netlink API for reading the routing
table, despite it being the only operating system supporting SIOCGETRPF.
On BSD we will continue to use the old routing socket API.  Hence, there
are no semantic changes introduced in this patch.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>