Clean up, bugfix and make new PIM Hello GenID feature the default
The new PIM Hello GenID feature, added in GIT commit 62dd428, only
added GenID to Hellos we sent.  It did not retrieve neighbor GenID
and use it to detect reboots.

This patch cleans up the GenID support and makes it the default.

On reception of a new GenID from a known neighbor we now also send
the neighbor the current RP-Set for faster recovery on reboot.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>