Fix issue #63: Corrected assert timer handling
It might be the case that the `MRTF_ASSERTED` flag was originally meant
to be used for inbound ASSERT detection only, but from now on we will be
using it for outbound handling as well.

With these changes ASSERT state is cleared after 180 seconds, hard coded
value, like before this fix.

Signed-off-by: Mika Joutsenvirta <mika.joutsenvirta@insta.fi>
Signed-off-by: Joachim Nilsson <troglobit@gmail.com>