debug: rate-limit only what is being logged
Currently logit() rate-limits everything but debug and serious messages,
not taking the --loglevel setting into account. This has the side-effect
that messages that would not be printed by syslog() due to the severity
filter, cause higher-severity messages to be suppressed for 10 minutes.

We fix this by taking syslog_level into account when bumping log_nmsgs.