Allow use of loopback addresses (/32)
pimd had to be compiled with --disable-masklen-check to allow the use of
loopback addresses (/32). The option is a bit a misnomer as it really
only enables the use of a /32, not any masklen.

As it is considered a good practice when having multiple interfaces to
put C-BSR/C-RP on a loopback interface, it seems better to enable the
ability to do so by default in pimd.

This commit just removes the compile-time option and enables it by
default.