Fix #68: Do not enumerate a VIF for disabled interfaces
Creating VIFs for all interfaces by default can be disabled using the
command line option `-N`.  However, pimd interanally still enumerated
VIFs used for the kernel, so even though interfaces were skipped they
stole the VIF n:o for enabled interfaces.

This patch adds a new pass to the pimd.conf file that runs before the
config_vifs_from_kernel() function.  When the latter is called it can
now query the configuration before even considering an interface.

TODO: The uvifs[] array, and the habit of passing around `vifi` (VIF
index) everywhere, is the root cause of our problems.  It would be far
better to keep a linked list, or something, of all interfaces and have
the VIF be allocated when installing it into the kernel.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>