Fix an issue on RP hash calculations.
As per RFC4601, all multicast routers in a domain should use the same
algorithm for RP elections, and all the IP addresses including the mask
address used in the algorithm should be in host byte order. The change fixes
the problem about RP hash calculation so that the results will be consistent
across different routers.

Signed-off-by: Xiaodong Xu <stid.smth@gmail.com>