Remove deprecated pimd.conf syntax from parser
With the transition to pimd v3.0 the following (already) deprecated
settings will no longer be available:

- cand_bootstrap_router
- cand_rp
- rp_address
- switch_register_threshold
- switch_data_threshold
- spt_threshold
- default_source_metric
- default_source_preference
- default_igmp_query_interval
- default_igmp_querier_timeout
- hello_period
- switch_register_threshold
- switch_data_threshold

Not only is it time to clear this out, it will also be a good wakeup
call to anyone who upgrades only parts of the PIM domain.

Remember, pimd v3 may not be compatible with older pimd releases!

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>