Fix #98: Restore line if token "masklen" does not match
When a phyint line with `scoped` does not contain a separate `masklen`
argument, but instead looks as follows, then the last argument `igmpv2`
is silently consumed by the `scoped` sub-parser.

    phyint eth0 scoped 224.0.1.0/24 igmpv2

This patch uses the previous commit (74e1867) to reset the parser for
each such sub-parser if it fails to find its expected token.  Issue #98
details the issue and lists both `scoped` and `altnet` as problematic,
the fix covers both.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>