Fix possible array overrun found by Coverity Scan
The k_req_incoming() API, available both as a routing socket (BSD) and a
netlink (Linux) back-end, was inconsistently used.  The callee assumed
that the API set rpf->iif to NO_VIF as a means to signal error, but the
Linux back-end failed to honor this, defaulting to ALL_VIFS instead.

The inconsistency led to possible -1 indexing of the uvifs[] array.

This patch refactors the k_req_incoming() API, so it shares defaults
between all back-ends.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>