Fix a double-free when daemon is restarting.
When pimd receives SIGHUP signal and in the Source-specific multicast case, it
will restart and the sources list in the VIF table will be freed. However the
source node gets double freed due to an error in walking the list. The correct
pointer to the next node needs to be set.

Signed-off-by: Xiaodong Xu <stid.smth@gmail.com>