pimd.service: Use correct command line option to run in foreground
For proper monitoring by systemd we must run pimd as a foreground
process.  In earlier versions of pimd we used `-f` for this, but
from 3.0 we've changed the command line option to `-n`.

This patch also enables the use of syslog, which is what is to be
expected of a UNIX daemon.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>