igmp: Simplify handling of IGMP v1/v2 vs v3
This patch removes a limitation on SSM range multicast groups to only be
available for IGMP v3 end devices.  In most legacy settings, industry in
particular, the operators of the networks are rarely informed of RFC
limitations.  It is no small feat to just explain what multicast is and
how TTL values work.  There is also no point in having limitations like
these in pimd; it complicates the code and adds no value to the user.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>