Issue #137: Possibly fix runtime memory leak
This patch fixes an ordering issue in rp.c:add_rp_grp_entry().  The call
to add_cand_rp() returns a pointer which may be lost if the subsequent
for() loop returns.

None of the code related to add_cand_rp() is needed before the for()
loop so it should be safe to move it and thus also fix this potential
runtime memory leak, as reported in issue #137.

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>