Make pimd behave like a normal UNIX daemon wrt logging to syslog/TTY
This patch removes all the (now) superfluous logging to stderr when
running in debug mode, we log state changes and the user has the new
pimctl tool for diagnostics.  Only -n enables logging to stdout/stderr,
wheres syslog is enabled by default, and when -s is explicitly given.

Also fixes a regression in logging to the TTY introduced in bdf546c by
removing the non-portable LOG_PERROR flag to openlog().

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>