Exit on error, even if running in the foreground
When logit() was refactored, ages ago, a small regression crept in that
was triggered when pimd runs in the foreground.  Many code paths rely on
calls to logit() to terminate¹ pimd when message severity is LOG_ERR.

The regression caused programs running in the foreground to ignore this
and this patch fixes that regression.
___

¹ only time pimd should *not* terminate for these calls to logit() is
  when it's been built with configure --disable-exit-on-error

Signed-off-by: Joachim Nilsson <troglobit@gmail.com>