Allow IGMP reports from src 0.0.0.0, RFC3376 sec. 4.2.13
This patch enables IP_PKTINFO (Linux) on the igmp_socket, which is later
used in igmp_read() to figure out incoming ifindex of IGMP membership
reports.  The ifindex is used to look up the inbound VIF, on which the
group is registered.

The purpose of all this is to be able to support end devices that have
not yet got an IP address, e.g. DHCP, or proxy reporting IGMP snoopers.

Non-Linux systems are not supported (yet).  NetBSD support IP_RCVPKTINFO
and FreeBSD have IP_RECVIF that can be used to the same effect.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>