Fix init of register tunnel on Linux
On Linux we support multiple MRT and multiple PIM router instances.
When that is in effect the system needs a unique name for each PIM
register tunnel.

Originally the PIM register tunnel interface name on all systems was
called register_vif0.  Since 2015, commit 419bde7, the name on Linux
systems is pimreg, or pimregN, where N is the instance if N > 0.

This patch is an update to 419bde7, saving the name in the uvifs[]
array so both kernel and daemon use the same name.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>