Import tests from pimd-dense and mrouted
The three.sh and pod.sh tests fail miserably.  Currently unclear if it's
got something to do with establishing an RP or something else, e.g. if
the bidirectional traffic of mping confuses pimd (it shouldn't!).

Publishing as-is to help illustrate the fact that pimd in its current
state really isn't ready for a v3.0 release yet.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>