Drop --no-fallback handling, completely unintuitive
This patch removes the fallback handling for RP/BSR added in c99c7f9.

The resolution of issue #66 is one of the worst decisions I've probably
made as a maintainer.  At its core is a misguided sense of getting pimd
to be more stand-alone, like mrouted.  But that's no excuse.  Turns out
it is *really* difficult to explain that having no config file differs
from commenting out the rp-candidate and bsr-candidate from the config
file.  A few years later, looking back at it, not able to even explain
it to myself, it becomes clear that we should just revert it.

With PIM-SM you *have* to know about Rendez-vous Points and what role
they play in the protocol.  Having some magic fairy dust to "simplify"
this, that is not standardized, is sort of doomed for failure.  In the
end, the simplification is PIM-SSM, which does not require an RP.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>